Experimenting with flexbox layout in main article.
Trying to contain maximum article width without introducing any layout/aesthetic elements.  Using em units to measure maximum width is a pain where there isn't a parent with a normalized font-size.  The description element in this example is too wide and would need to have it's width set proportional to it's font-size.  Not workable.